# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Renta de Apartamentos - API",
        "description": "Documentación de la API del sistema de rentas de apartamentos. Permite gestionar usuarios, apartamentos, quejas, contratos y más.",
        "contact": {
            "email": "tucorreo@ejemplo.com"
        },
        "license": {
            "name": "Todos los derechos reservados - Universidad Piloto de Colombia"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rol-permiso": {
            "get": {
                "tags": [
                    "RolPermiso"
                ],
                "summary": "Listar todas las relaciones rol-permiso",
                "operationId": "db579b6f2064258859edf7408c08095d",
                "responses": {
                    "200": {
                        "description": "Lista de relaciones rol-permiso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RolPermiso"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "RolPermiso"
                ],
                "summary": "Crear una nueva relación rol-permiso",
                "operationId": "5603377bf037e5854c20289d8f8f9637",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rol_id",
                                    "permiso_id"
                                ],
                                "properties": {
                                    "rol_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "permiso_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Relación rol-permiso creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RolPermiso"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/rol-permiso/{id}": {
            "get": {
                "tags": [
                    "RolPermiso"
                ],
                "summary": "Obtener una relación rol-permiso específica",
                "operationId": "d94f47b3a79f18ed5d2a40874c7571c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Relación encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RolPermiso"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Relación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "RolPermiso"
                ],
                "summary": "Actualizar una relación rol-permiso existente",
                "operationId": "b9f3e06df428fd42188651ad2d38fc76",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rol_id",
                                    "permiso_id"
                                ],
                                "properties": {
                                    "rol_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "permiso_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Relación actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RolPermiso"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Relación no encontrada"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "RolPermiso"
                ],
                "summary": "Eliminar una relación rol-permiso",
                "operationId": "94707c0102beb31f5fe4cbe57ecca45b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Relación eliminada exitosamente"
                    },
                    "404": {
                        "description": "Relación no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RolPermiso": {
                "title": "RolPermiso",
                "description": "Relación entre un rol y un permiso",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "rol_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "permiso_id": {
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}